Fix javaclient testkit failures and align v6 conformance
The WIP checkpoint left 11 testkit failures when the suite was first
wired onto the new javaclient backend. This commit drives them to
zero, in the process exposing two real backend bugs and one v5 API
shape that doesn't fit v6.

Backend bugs fixed (uncovered by promoted contract tests):
  - configureBuilderOrRetrieve didn't check callbackState before
    metric registration, so cross-axis (metric vs callback) name
    collisions were silently allowed. Mirrors the v5 check.
  - registerCallback keyed StateKey under name.show (full counter
    name, e.g. foo_total) while metric registration keyed under
    counterName(name) (stripped, foo). Same-name collision detection
    missed counter cases. Aligned to the stripped form.
  - Error messages also surfaced the stripped form to users; added a
    renderedName parameter so error text uses the wire-level name
    (foo_total) like v5.

API shape change (Summary.Value.count: Double -> Long):
  - Upstream prometheus-metrics-core 1.x SummaryDataPointSnapshot
    constructor takes long count; v5 simpleclient took double.
    Summary.Value previously typed count as the value-type parameter A
    (Double or Long); v6 fixes count to Long for both
    Value[Double] and Value[Long]. Counts are integer in Prometheus.
    Source-incompatible at every Summary.Value(...) construction site
    (compile errors, no silent breaks). Trivial mechanical fix at
    each call site (1.0 -> 1L).

Other v6 behaviour alignments:
  - createAndRegisterInfo registered under name.value (no prefix);
    snapshot lookups via NameUtils.makeName(prefix, name) failed to
    find it. Corrected to register under the prefix-qualified name.
  - registerDoubleCounterCallback / metricCollectionToSnapshots clamp
    negative values to 0 (matches v5 ergonomics; v6 upstream
    CounterDataPointSnapshot.validate() throws on negative).
  - getExemplarHistogramValue testkit helper: exemplar bucket
    assignment is per-bucket-range (lower < value <= upper), not
    cumulative. Previously cumulative matching attached the same
    exemplar to every bucket >= the value.

Testkit assertion updates (consistent contract for both backends):
  - Counter / metric-collection tests use a non-negative round-trip
    expectation (clamped to 0 if user supplied negative). Documents
    the v6 stricter semantics; v5 incidentally satisfied this where
    its callback paths already clamped.
  - getSummaryValue returns Option[Long] for count.

Five name-collision contract tests promoted from
javasimpleclient.JavaMetricRegistrySuite into the new javaclient
JVM-specific suite (kept here, not in the cross-backend testkit,
because they need a single registry instance implementing both
MetricRegistry and CallbackRegistry; the testkit's separate
resources would yield disjoint internal state and defeat the
cross-axis collision check).

Test count: javaclient testkit + JVM-specific = 38 + 19 = 57 (all
green). javasimpleclient still has 47 testkit tests; 2 of the
summary callback tests fail because Long counts > 2^53 lose
precision through simpleclient's Double-typed SummaryMetricFamily
wire format. The legacy backend is removed in the following commit;
the failures disappear with it.

+++ modules/prometheus4cats-java/src/main/scala/prometheus4cats/javasimpleclient/JavaMetricRegistry.scala
@@ -690,7 +690,7 @@ class JavaMetricRegistry[F[_]: Async] private (
       f,
       (n, lns, lvs, v) =>
         if (v.quantiles.isEmpty)
-          new SummaryMetricFamily(n, help.value, lns.asJava).addMetric(lvs.asJava, v.count, v.sum)
+          new SummaryMetricFamily(n, help.value, lns.asJava).addMetric(lvs.asJava, v.count.toDouble, v.sum)
         else
           new SummaryMetricFamily(
             n,
@@ -700,7 +700,7 @@ class JavaMetricRegistry[F[_]: Async] private (
           )
             .addMetric(
               lvs.asJava,
-              v.count,
+              v.count.toDouble,
               v.sum,
               v.quantiles.values.toList.map(_.asInstanceOf[java.lang.Double]).asJava
             )
